month06/week23/session69/products-fullstack/backend/requirements.txt
asgiref==3.11.0
Django==5.2.9
django-cors-headers==4.9.0
djangorestframework==3.16.1
pillow==12.0.0
psycopg2-binary==2.9.11
sqlparse==0.5.4
